+++ Assets/Tests/InputSystem/CoreTests_Controls.cs
@@ -439,7 +439,7 @@ partial class CoreTests
 
         var receivedCalls = 0;
         InputSystem.onEvent +=
-            eventPtr =>
+            (eventPtr, device) =>
         {
             ++receivedCalls;
             float value;
@@ -479,11 +479,10 @@ partial class CoreTests
 
         float? value = null;
         InputSystem.onEvent +=
-            eventPtr =>
+            (eventPtr, _) =>
         {
             Assert.That(value, Is.Null);
-            float eventValue;
-            ((AxisControl)device["extraControl"]).ReadValueFromEvent(eventPtr, out eventValue);
+            ((AxisControl)device["extraControl"]).ReadValueFromEvent(eventPtr, out var eventValue);
             value = eventValue;
         };
 
@@ -502,7 +501,7 @@ partial class CoreTests
 
         var receivedCalls = 0;
         InputSystem.onEvent +=
-            eventPtr =>
+            (eventPtr, device) =>
         {
             ++receivedCalls;
             gamepad.leftTrigger.WriteValueIntoEvent(0.1234f, eventPtr);
@@ -523,7 +522,7 @@ partial class CoreTests
 
         var receivedCalls = 0;
         InputSystem.onEvent +=
-            eventPtr =>
+            (eventPtr, device) =>
         {
             ++receivedCalls;
             gamepad.leftTrigger.WriteValueIntoEvent(0.1234f, eventPtr);
@@ -1137,4 +1136,44 @@ partial class CoreTests
         Assert.That(UnsafeUtility.SizeOf<TouchState>(), Is.EqualTo(TouchState.kSizeInBytes));
         Assert.That(touchscreen.touches[0].stateBlock.alignedSizeInBytes, Is.EqualTo(TouchState.kSizeInBytes));
     }
+
+    #if UNITY_EDITOR || DEVELOPMENT_BUILD
+    [Test]
+    [Category("Controls")]
+    public void Controls_CanAddDebugVisualizerForControl()
+    {
+        var gamepad1 = InputSystem.AddDevice<Gamepad>();
+        InputSystem.AddDevice<Mouse>(); // Noise.
+
+        var go = new GameObject();
+        go.SetActive(false);
+        var visualizer = go.AddComponent<InputControlVisualizer>();
+        visualizer.controlPath = "<Gamepad>/buttonSouth";
+        go.SetActive(true);
+
+        Assert.That(visualizer.control, Is.SameAs(gamepad1.buttonSouth));
+
+        InputSystem.RemoveDevice(gamepad1);
+
+        Assert.That(visualizer.control, Is.Null);
+
+        InputSystem.AddDevice(gamepad1);
+        var gamepad2 = InputSystem.AddDevice<Gamepad>();
+
+        Assert.That(visualizer.control, Is.SameAs(gamepad1.buttonSouth));
+
+        visualizer.controlIndex = 1;
+
+        Assert.That(visualizer.control, Is.SameAs(gamepad2.buttonSouth));
+
+        visualizer.controlIndex = 2;
+
+        Assert.That(visualizer.control, Is.SameAs(gamepad1.buttonSouth));
+
+        go.SetActive(false);
+
+        Assert.That(visualizer.control, Is.Null);
+    }
+
+    #endif // UNITY_EDITOR || DEVELOPMENT_BUILD
 }